Added custom AlertDialog with spinner for new projects + bugfix
Spinner gets scaffoldingSystems from Firebase. Bugfix for walls without pictures couldnt upload pictures

+++ app/src/main/java/no/hiof/bo20_g28/stillashjelpen/WallActivity.java
@@ -276,9 +276,11 @@ public class WallActivity extends AppCompatActivity {
     }
 
     private void deleteWallsOldPicture(){
-        StorageReference storageReference = FirebaseStorage.getInstance().getReference().child("images");
-        StorageReference imageReference = storageReference.child(thisWall.getPictureId());
-        imageReference.delete();
+        if(thisWall.getPictureId() != null) {
+            StorageReference storageReference = FirebaseStorage.getInstance().getReference().child("images");
+            StorageReference imageReference = storageReference.child(thisWall.getPictureId());
+            imageReference.delete();
+        }
     }
